pyproject.toml
[project]
dependencies = [
    "asyncer>=0.0.8",
    "dateparser>=1.2.1",
    "feedparser>=6.0.11",
    "loguru>=0.7.3",
    "pyrotgfork[fast]>=2.2.12",
    "pysocks>=1.7.1",
    "python-ffmpeg>=2.0.12",
    "requests>=2.32.3",
    "typer>=0.16.0",
    "yt-dlp>=2025.1.12rc",
]
name = "tubesync"
requires-python = "~=3.12"
version = "0.1.0"

[dependency-groups]
dev = ["ipdb>=0.13.13"]

# https://docs.astral.sh/ruff/settings
[tool.ruff]
cache-dir = '~/.cache/ruff'
force-exclude = true
indent-width = 4
line-length = 200
output-format = 'grouped'
respect-gitignore = true
show-fixes = true
target-version = 'py312'

[tool.ruff.format]
indent-style = 'space'
line-ending = 'lf'
quote-style = 'double'

[tool.ruff.lint]
ignore = ['ANN', 'PTH', 'T20', 'ERA', 'C90', 'D1', 'INP001', 'E203', 'E266', 'E501', 'EXE001', 'PD901', 'PGH003', 'S101', 'S301', 'UP009', 'COM812', 'ISC001', 'PLR2004', 'S311']
ignore-init-module-imports = true
select = ['ALL']

[tool.ruff.lint.pydocstyle]
convention = 'google'

# https://lyz-code.github.io/yamlfix/#configuration
[tool.yamlfix]
allow_duplicate_keys = false
comments_min_spaces_from_content = 2
comments_require_starting_space = true
comments_whitelines = 1
explicit_start = 1
indent_mapping = 2
indent_offset = 2
indent_sequence = 4
line_length = 200
none_representation = ''
preserve_quotes = true
quote_basic_values = false
quote_keys_and_basic_values = false
quote_representation = '"'
section_whitelines = 1
sequence_style = 'keep_style'
whitelines = 1
